restructuring entity package to be base model with request and entity as sub packages
