+++ src/main/java/com/groupthree/bankapp/model/entity/Account.java
@@ -1,11 +1,8 @@
-package com.groupthree.bankapp.entity;
+package com.groupthree.bankapp.model.entity;
 
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
-import com.groupthree.bankapp.entity.CheckingAccount;
-import com.groupthree.bankapp.entity.InterestAccount;
-import com.groupthree.bankapp.entity.RegularAccount;
-import org.hibernate.annotations.GenericGenerator;
+import com.groupthree.bankapp.exception.AccountTypeNotSupportedException;
 
 import javax.persistence.*;
 
@@ -42,6 +39,42 @@ public abstract class Account {
     private double interestCharge;
 
     private double amount;
+
+    /**
+     * Factory method to create new instances of the Account child classes
+     * based on the type provided
+     *
+     * @param accType The account type to create. Supports: regular, checking, interest
+     * @param name The name to register to the account under
+     * @param accountNumber The account number to register the account under
+     * @return The created account
+     * @throws AccountTypeNotSupportedException If the account type provided is not supported
+     */
+    public static Account of(String accType, String name, String accountNumber) throws AccountTypeNotSupportedException {
+        Account newAcc;
+        switch (accType) {
+            case "regular" : {
+                newAcc = new RegularAccount();
+                break;
+            }
+            case "checking" : {
+                newAcc = new CheckingAccount();
+                break;
+            }
+            case "interest" : {
+                newAcc = new InterestAccount();
+                break;
+            }
+            default: {
+                throw new AccountTypeNotSupportedException();
+            }
+        }
+
+        newAcc.setName(name);
+        newAcc.setAcctNumber(accountNumber);
+
+        return newAcc;
+    }
 
     public Long getId() {
         return id;
