+++ src/main/java/com/groupthree/bankapp/model/entity/CheckingAccount.java
@@ -1,21 +1,24 @@
-package com.groupthree.bankapp.entity;
+package com.groupthree.bankapp.model.entity;
 
 import javax.persistence.Entity;
 
 @Entity
-public class RegularAccount extends MinimumBalanceAccount {
-    public RegularAccount() {
+public class CheckingAccount extends MinimumBalanceAccount {
+    public CheckingAccount() {
         super();
-        setBalance(500);
-        setMinimumBalance(500);
+        setBalance(100);
+        setMinimumBalance(100);
         setInterestCharge(0);
+        setTransactionCharge(1);
         setPenalty(10);
     }
 
     @Override
     public void setBalance(double balance) {
         balance = this.verifyWithinMinimum(balance);
+        balance -= getTransactionCharge();
         this.balance = balance;
         this.verifyOverMinimum();
     }
+
 }
