+++ src/main/java/com/groupthree/bankapp/model/entity/MinimumBalanceAccount.java
@@ -1,12 +1,10 @@
-package com.groupthree.bankapp.entity;
+package com.groupthree.bankapp.model.entity;
 
 public abstract class MinimumBalanceAccount extends Account {
 
     protected MinimumBalanceAccount(){
         setDidPassMinimumBalance(false);
     }
-
-
 
     public double verifyWithinMinimum(double balance){
         // penalty should not be charged after falling below min balance once
