+++ src/main/java/com/groupthree/bankapp/model/entity/RegularAccount.java
@@ -1,24 +1,21 @@
-package com.groupthree.bankapp.entity;
+package com.groupthree.bankapp.model.entity;
 
 import javax.persistence.Entity;
 
 @Entity
-public class CheckingAccount extends MinimumBalanceAccount {
-    public CheckingAccount() {
+public class RegularAccount extends MinimumBalanceAccount {
+    public RegularAccount() {
         super();
-        setBalance(100);
-        setMinimumBalance(100);
+        setBalance(500);
+        setMinimumBalance(500);
         setInterestCharge(0);
-        setTransactionCharge(1);
         setPenalty(10);
     }
 
     @Override
     public void setBalance(double balance) {
         balance = this.verifyWithinMinimum(balance);
-        balance -= getTransactionCharge();
         this.balance = balance;
         this.verifyOverMinimum();
     }
-
 }
